requirements.txt
anthropic
flask
flask-cors
requests
python-dotenv
